Fix where go tarball is extracted to by default
For some reason I got this mixed up for another extraction path. The
default extraction path makes more sense to be /usr/local as this is
what go.dev uses in their example.

Though this said, it should also be modified according to the
individual project that is invoking the role. Hence, this variable can
be "considered" public.

Also, add TODOs for the golang tasks.