Initial commit to add jenkins containers
Should note this is for jenkins containers that rely on the jenkins-base
container. For development and deployment purposes, I wanted to have
a common way to instantiate each image like that of using docker
compose.

The added benefit being, I only have to change the "compose" like
definition in one place. Rather than in multiple places (e.g. in other
playbooks and makefiles that assist in development of the containers
through testing).