Discard changed_when when dir(s) are absent
The state attribute is set on the registered variable based on the
task module's state parameter. If the state is set to directory, then
the state should never be "absent". However I am unsure why I did
this in the first place, as the file module will not report a changed
status if the file is in the state defined.