Handle idempotence for lxd systemd service file
I noticed when the LXD systemd service file changed and was copied over
to a system that already had LXD running, that, the systemd daemon was
not automatically reloaded to account for the new service file. This was
done and additionally I needed to have this handler flushed after the
new service file was copied over.

This however conflicts with the original intent of the placement of
saving the iptables rules. I believe this handler was meant to run after
the LXD service is running but currently this appears to run after the
main tasklist finishes. Which I'm ok with. So for now this handler will
only be notified if the service starts, which will be the case on a
fresh install or if the service file for systemd/sysvinit change.