Look into runtime ansible version vs role version
It appears the runtime ansible version is actually a PEP440 version
specifier, meaning, it may also contain verison clause(s). Hence this
might be why the requires_ansible is requires its value be in quotes.

As for making things consistent in the roles's min ansible version. I
will leave things as because other projects do not quote their versions
of software. General rule of thumb is not to use quotes unless I need
to.