Enable replacing go version & install for any user
Now when using the golang role, go can be installed for any user
depending on how the calling Ansible code runs the role (e.g.
ansible_user or by become).

Now the golang role will replace the go version it finds for a
particular user, assuming the requested golang version is different than
what is detected.

I've decided to let each individual project worry about the GOPATH to be
set when using go to build what it needs. The idea also being perhaps
the binaries being build should not be globally accessible (e.g.
binaries are put in a directory that is part of the PATH).