Remove golang role tmp dir, & previous go bin dir
Now any go tarball will be donwloaded directly into a temporary
directory to be extracted. This mainly because I do not see a need a
reason to dedicate an entire directory to one tarball being downloaded.
Hence I just let the tarball be downloaded into a temporary directory.
To add, the go tarball will only be downloaded if the go version is
different than what the user can detect is currently on the PATH.

When a new version of go is installed, the previous PATH set with the go
binary will be removed before adding the new setting of the PATH with
the new go version.