Remove the community.general collection dependency
This is because this collection is already part of the "ansible" package
and as of this commit, there isn't an exact version needed of this
collection for cavcrosby.general to depend on. If the community.general
collection is removed from the standard ansible package, then this will
need to revisited and added back.

That said, by locking the version of ansible used, this effectively
should lock the version of this collection until the ansible package
version is updated.