Rename env var from 1c6f90d4
Part of the environment variable name was from the no_log task
parameter. This makes its intended purpose somewhat harder to realize
when setting the value for this environment variable (as I ran into this
tonight).

The environment variable name should now reflect more clearly whether to
log jenkins-base-container (jbc) secrets or not.