Correct 77d04dde to do what user expects
The environment variable, while easier to think about its intended
purpose, would actually do the opposite of what's intended.

The configuration code has been corrected to log secrets if the
environment variable is set with a "truthy" value. By default (or if the
environment variable is not set), secrets will not be logged.