Refactor Makefile
collection-meta doesn't exist as a target, hence removing the Makefile
variable designated for it.

The ansible-lint target is now just called 'lint' because that's the
generic target name I use in other projects.

The order of ansible-galaxy options has been rearranged as the order of
command parameters should be as followed: options, options with option
arguments, and finally positional arguments to the command.

Makefile recipes have been changed to match my style of shell scripting
(e.g. splitting of long lines where possible).